+++ Tests/Resources/Fixtures/config/full.php
@@ -13,6 +13,15 @@ $container->loadFromExtension('httplug', [
         'uri_factory'     => 'Http\Message\UriFactory\GuzzleUriFactory',
         'stream_factory'  => 'Http\Message\StreamFactory\GuzzleStreamFactory',
     ],
+    'clients' => [
+        'test' => [
+            'factory' => 'httplug.factory.guzzle6',
+            'http_methods_client' => true,
+            'options' => [
+                'default_host' => 'http://localhost',
+            ],
+        ],
+    ],
     'profiling' => [
         'enabled' => true,
         'formatter' => 'my_toolbar_formatter',